feat: move around interaction processing
- made database initialization to use a provided URL or default to 'database.sqlite'
- Better rate limiting in emoji creation
- Refactored interaction handling to support various interaction types (commands, buttons, modals, select menus) without conflicting event handling
- Added error handling and feedback for interaction errors.
- Removed panelBuilder and wizardSetup event files.
- Added input validation for panel title and description lengths.
