+++ src/utils/panel/panelDetails.js
@@ -4,6 +4,7 @@ import {
   ButtonStyle,
   EmbedBuilder,
   Message,
+  MessageFlags,
   ModalBuilder,
   TextInputBuilder,
   TextInputStyle,
@@ -54,6 +55,24 @@ export const handleDetailsSubmission = async (
   const panelDescription = interaction.fields.getTextInputValue(
     'panel_description',
   );
+
+  // validate input lengths
+  if (panelTitle.length > 256) {
+    await interaction.followUp({
+      content: 'Panel title must be 256 characters or less.',
+      flags: MessageFlags.Ephemeral,
+    });
+    return;
+  }
+
+  if (panelDescription.length > 4096) {
+    await interaction.followUp({
+      content: 'Panel description must be 4096 characters or less.',
+      flags: MessageFlags.Ephemeral,
+    });
+    return;
+  }
+
   const msg = interaction.message;
 
   const firstEmbed = EmbedBuilder.from(msg.embeds[0])
